backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-allauth>=0.57.0
dj-rest-auth>=4.0.0
django-cors-headers>=4.3.1
requests>=2.31.0
python-dotenv>=1.0.0 